% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dummy.R
\name{as_dummy}
\alias{as_dummy}
\title{Convert Factor Variables to Dummies}
\usage{
as_dummy(x, term, factor = F, labels = c("Off", "On"))
}
\arguments{
\item{x}{A vector of factor data to convert to a dummy.}

\item{term}{A vector of terms to recode as 1.}

\item{factor}{Convert the dummy to a factor variable? Defaults to FALSE.}

\item{labels}{If factor = T, a character vector to specify the names of the resulting levels (e.g. c("Tails", "Heads")). Defaults to c("Off", "On").}
}
\value{
A vector of dummy data.
}
\description{
This function takes away the need to rely on ifelse() to create dummy variables.
}
\examples{
x <- sample(c("Coffee", "Tea", "Hot Chocolate"), replace = T, size = 100)
as_dummy(x, term = "Coffee")
}
