% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bailey_colours.R
\name{bailey_colours}
\alias{bailey_colours}
\title{Jack Bailey's Colour Palette.}
\usage{
bailey_colours(...)
}
\description{
This is a convenience function that makes it easier for me to make plots that use colours I like.
}
