% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_swing.R
\name{compute_swing}
\alias{compute_swing}
\title{Compute Two-Party Swing}
\usage{
compute_swing(party_a = NULL, party_b = NULL)
}
\arguments{
\item{party_a}{A numeric vector of length 2. The first value corresponds to the most recent election.}

\item{party_b}{A numeric vector of length 2. The first value corresponds to the most recent election.}
}
\description{
This convenience function takes two numeric vectors of vote share data and returns the estimated two-party swing.
}
\examples{
\dontrun{
compute_swing(party_a = c(0.5, 0.4), party_b = c(0.5, 0.6))
}
}
