% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse Logit Function}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{A vector of numeric data.}
}
\description{
Convert variables from the logit to probability scale. A port of the same function from the rethinking package.
}
\examples{
x <- rnorm(100)
inv_logit(x)
}
