\name{is.vector2}
\alias{is.vector2}
\title{Checks safely if a given object is a vector.
(\code{base::is.vector} returns \code{TRUE} only if the given vector
has no attributes beside \code{names()}).}
\usage{
is.vector2(x)
}
\arguments{
  \item{x}{[\code{mixed}]\cr Object to check.}
}
\value{
[\code{logical(1)}] Logical value indicating whether the
given object is a vector or not.
}
\description{
Checks safely if a given object is a vector.
(\code{base::is.vector} returns \code{TRUE} only if the
given vector has no attributes beside \code{names()}).
}

