% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_objects.R
\name{list_objects}
\alias{list_objects}
\title{Create List of All Objects in the Global Environment}
\usage{
list_objects()
}
\value{
A list object.
}
\description{
This function returns a list of all objects currently stored in the global environemnt. Useful for saving the output of an analysis in one go.
}
\examples{
list_objects()
}
