% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorem.R
\name{lorem}
\alias{lorem}
\title{Generate Placeholder Text}
\usage{
lorem(length = 1)
}
\arguments{
\item{length}{The length of the resulting character string.}
}
\value{
A character string.
}
\description{
This function generates placeholder text ("Lorem ipsum...") as needed.
}
\examples{
lorem(length = 11)
}
