% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_na.R
\name{mark_na}
\alias{mark_na}
\title{Mark Values as NA}
\usage{
mark_na(x, term, drop = T)
}
\arguments{
\item{x}{A vector of factor data to convert to a dummy.}

\item{term}{A vector of terms to recode as 1.}

\item{drop}{Drop missing levels? Defaults to TRUE.}
}
\value{
A vector of data, with missing data marked as NA.
}
\description{
This function makes it easier to convert values to missing (NA).
}
\examples{
x <- sample(c("Coffee", "Tea", "Hot Chocolate"), replace = TRUE, size = 100)
mark_na(x, term = "Coffee")
}
