% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncr.R
\name{ncr}
\alias{ncr}
\title{Compute Combinations}
\usage{
ncr(n = 1, r = 1)
}
\arguments{
\item{n}{Number of sample points in set n.}

\item{r}{Number of sample points in each combination.}
}
\value{
A numbers.
}
\description{
Compute arbitrary combinations (order doesn't matter).
}
\examples{
ncr(n = 5, r = 2)
}
