% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npr.R
\name{npr}
\alias{npr}
\title{Compute Permutations}
\usage{
npr(n = 1, r = 1)
}
\arguments{
\item{n}{Number of sample points in set n.}

\item{r}{Number of sample points in each permutation.}
}
\value{
A numbers.
}
\description{
Compute arbitrary permutations (order matters).
}
\examples{
npr(n = 5, r = 2)
}
