\name{permutate}
\alias{permutate}
\title{Simple wrapper function for sample.}
\usage{
permutate(x)
}
\arguments{
  \item{x}{[\code{vector}]\cr Source vector.}
}
\value{
[\code{vector}] Source vector with permutated elements.
}
\description{
One more function call necessary, but as a result the code
is much more read friendly.
}

