% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale Variables in R}
\usage{
rescale(x, type = "z", na.rm = T)
}
\arguments{
\item{x}{A vector of interval-type data to standardise.}

\item{type}{One of "z", "mean", "gelman". Option "z" converts the scale to z-scores. Option "mean" rescales the data so that its mean is zero. Option "gelman" rescales the data so that it is centred on zero and a one-unit increase or decrease corresponds with the scale moving to its 97.5th and 2.5th percentile.}

\item{na.rm}{Remove missing data? Defaults to TRUE.}
}
\value{
A vector of rescaled data.
}
\description{
This functions makes it easier to rescale continuous variables.
}
\examples{
x <- rnorm(100, 1, .2)
rescale(x, type = "z")
}
