% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlikert.R
\name{rlikert}
\alias{rlikert}
\title{Simulate Likert-Type Data}
\usage{
rlikert(
  n = 1000,
  t1 = -0.84,
  t2 = -0.25,
  t3 = 0.25,
  t4 = 0.84,
  mean = 0,
  disc = 0
)
}
\arguments{
\item{n}{The number of random outcomes to generate.}

\item{t1}{The first threshold parameter. Defaults to -0.84.}

\item{t2}{The second threshold parameter. Defaults to -0.25.}

\item{t3}{The third threshold parameter. Defaults to 0.25.}

\item{t4}{The fourth threshold parameter. Defaults to 0.84.}

\item{mean}{The mean of the latent continuous variable that underpins the ordinal outcome. Defaults to 0.}

\item{disc}{The discrimination of the latent continuous variable that underpins the ordinal outcome. Defaults to 0.}
}
\value{
A vector of likert type data.
}
\description{
Simulate ordinal variables of arbitrary length.
}
\examples{
rlikert(n = 10)
}
