\name{strExplode}
\alias{strExplode}
\title{Split a string into substrings.}
\usage{
strExplode(x, sep = " ")
}
\arguments{
  \item{x}{[\code{character}]\cr Source string.}

  \item{sep}{[\code{character}]\cr Seperator whcih is used
  to split \code{x} into substrings.}
}
\value{
[\code{character}] Vector of substrings.
}
\description{
A seperator determines which characters to use for spliting
up.
}

