\name{strImplode}
\alias{strImplode}
\title{Glue a vector of characters together.}
\usage{
strImplode(x, sep = " ")
}
\arguments{
  \item{x}{[\code{arbitrary}]\cr Arbitrary R object
  (converted into character before glueing).}

  \item{sep}{[\code{character(1)}]\cr Seperator used to
  glue elements together.}
}
\value{
[\code{character(1)}]
}
\description{
Glue a vector of characters together.
}

