% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bailey.R
\name{theme_bailey}
\alias{theme_bailey}
\title{Jack Bailey's Custom ggplot Theme.}
\usage{
theme_bailey(rmd = T)
}
\description{
This is a convenience function that makes it easier for me to make nice plots. It is an extension of theme_minimal() in ggplot2. It removes some lines, adds others, and changes the font to Cabin.
}
