% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uom_colours.R
\name{uom_colours}
\alias{uom_colours}
\title{University of Manchester Colour Palette.}
\usage{
uom_colours(...)
}
\description{
This is a convenience function that makes it easier for me to make plots that conform to the University of Manchester's corporate colour scheme.
}
