\name{usedFactors}
\alias{usedFactors}
\title{Determines the factor levels of a factor type vector
that are actually occuring in a given vector.}
\usage{
usedFactors(x)
}
\arguments{
  \item{x}{[\code{vector}]\cr Source vector of factors.}
}
\value{
[\code{character}] Character vector of used levels.
}
\description{
Determines the factor levels of a factor type vector that
are actually occuring in a given vector.
}

