% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlikert.R
\name{rlikert}
\alias{rlikert}
\title{Simulate Likert-Type Data}
\usage{
rlikert(n = 100, length = 5, mean = 0, disc = 0, sd = NULL)
}
\arguments{
\item{n}{The number of random outcomes to generate.}

\item{length}{The number of response options.}

\item{mean}{The mean of the latent continuous variable that underpins the ordinal outcome. Defaults to 0.}

\item{disc}{The discrimination of the latent continuous variable that underpins the ordinal outcome. Defaults to 0.}

\item{sd}{The standard deviation of the latent continuous variable that underpins the ordinal outcome. Defaults to NULL.}
}
\value{
A vector of likert type data.
}
\description{
Simulate ordinal variables of arbitrary length.
}
\examples{
rlikert(n = 10)
}
