% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_n_of.R
\name{sample_n_of}
\alias{sample_n_of}
\title{Sample Groups at Random}
\usage{
sample_n_of(data, size, replace = TRUE, ...)
}
\arguments{
\item{data}{A tibble or data frame.}

\item{size}{The number of groups you want to sample.}

\item{replace}{Whether to sample with replacement or not. Defaults to TRUE.}

\item{...}{The group you want to sample.}
}
\description{
Sample groups from a data set just as one would do using sample_n(). Based on a similar function from TJ Mahr's package tjmisc.
}
\examples{
iris \%>\%
sample_n_of(1000, )
}
